pyproject.toml
[project]
name = "FSM"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "HEROgold", email = "martijnwieringa28@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.5",
    "confkit>=0.3.0",
    "cryptography>=43.0.1",
    "flask[async]>=3.1.2",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "sqlalchemy>=2.0.43",
    "docker>=7.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
